move component webbroweserinput
