+++ WebApp/client/public/videoplayer/js/register-events.js
@@ -1,6 +1,6 @@
 import { gamepadHandler } from "./gamepadEvents.js";
-import * as Logger from "./logger.js";
-import { Keymap } from "./keymap.js";
+import * as Logger from "../../module/logger.js";
+import { Keymap } from "../../module/keymap.js";
 
 const InputEvent = {
   Keyboard: 0,
